requirements.txt
streamlit
psycopg2-binary
pandas
plotly
altair
python-dotenv
prophet
httpx
loguru
plotly
scikit-learn
xgboost
lightgbm
joblib
jupyterlab
notebook
SQLAlchemy==2.0.35